Moved File System Header Library
Moved header library `includes/filesys` to `includes/io/filesys`.
Renamed class `Scanner` to `TcpScanner`.
Other small changes like refactoring and reformatting.
Recompiled executable binaries.

+++ SvcScan/src/file_stream.cpp
@@ -3,7 +3,7 @@
 *  ---------------
 *  Source file for a system file stream
 */
-#include "includes/filesys/file_stream.h"
+#include "includes/io/filesys/file_stream.h"
 
 /**
 * @brief  Initialize the object.
@@ -172,7 +172,7 @@ std::string scan::FileStream::read_text(const bool &t_close)
     // Read the file data
     if (fsize != INVALID_SIZE)
     {
-        fdata = string(static_cast<uint>(fsize), '\0');
+        fdata = string(static_cast<size_t>(fsize), '\0');
         m_file.rdbuf()->sgetn(&fdata[0], fsize);
     }
